+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.0.8
 
--
+- convert to case: camelCase,pascalCase,constantCase,dotCase,kebabCase,snakeCase,pathCase,titleCase,sentenceCase,lowerCase,upperCase；
 
 ## 2.0.7